temp: plugin API
